requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
nltk==3.8.1
scikit-learn==1.4.1
wordcloud==1.9.3
mysql-connector-python==8.3.0
matplotlib==3.8.3
seaborn==0.13.2 